deploy/control-center/backend/pyproject.toml
[project]
name = "omnibioai-control-center"
version = "0.1.0"
description = "OmniBioAI Control Center - Stateless system health dashboard for the local stack"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
